fix(events): create api history page with START_API, STOP_API events and deploy the last published API if event is START or STOP
