+++ src/app/api/admin/events/_apiEvents.scss
@@ -1,76 +1,18 @@
 .api-events-timeline {
-  overflow-y: auto;
-  height: calc(100vh - 200px);
-}
+  .timeline:before {
+    width: 1px;
+  }
 
-.timeline li timeline-badge {
-  z-index: 0 !important;
-}
-
-.timeline li timeline-panel {
-  padding: 7px !important;
-}
-
-.timeline li timeline-panel:before {
-  top: 25px !important;
-}
-
-.timeline li timeline-panel:after {
-  top: 26px !important;
-}
-
-.timeline timeline-panel.active {
-  background-color: #ededed;
-}
-
-.timeline timeline-panel.active:after {
-  border-right: none;
-}
-
-.timeline .timeline-link {
-  color: #333;
-}
-
-.timeline:before {
-  position: absolute;
-  width: 1px;
-  left: 6%;
-}
-
-.timeline li timeline-badge {
-  box-shadow: none;
-  width: 20px;
-  height: 20px;
-  line-height: 20px;
-  font-size: 1em;
-  left: 13%;
-  top: 29px;
-}
-
-.timeline li.timeline-inverted timeline-panel {
-  float:left;
-  left: 18%;
-}
-
-ul.timeline li timeline-panel {
-  width: 80%;
-}
-
-pre.api-events-definition {
-  background-color: transparent !important;
-  border: none !important;
-}
-
-.api-events-checkbox {
-  position: absolute;
-  margin-bottom: 0px;
-  top: -2px;
-  margin-top: 15px;
-  right: 0px;
-}
-
-.api-events-icon {
-  position: absolute;
-  margin-top: 5px;
-  right: 9px;
+  .timeline li timeline-badge {
+    z-index: 0;
+    box-shadow: none;
+    height: 40px;
+    width: 40px;
+    line-height: 40px;
+    left: 51%;
+  }
+  
+  .timeline li timeline-panel {
+    padding: 7px;
+  }
 }
