+++ src/app/api/admin/events/apiEvents.html
@@ -15,51 +15,24 @@
     limitations under the License.
 
 -->
-<div layout="row" flex class="api-events" style="margin-top: 20px;">
-
-  <div class="api-events-timeline" flex="20">
+<div layout="row" layout-align="center center" flex style="margin-top: 20px;">
+  <div class="api-events-timeline" flex="50">
     <timeline>
       <div ng-repeat="eventTimeline in apiEventsCtrl.eventsTimeline">
-        <timeline-event side="right">
+        <timeline-event>
           <timeline-badge class="{{eventTimeline.badgeClass}}">
             <i class="glyphicon {{eventTimeline.badgeIconClass}}"></i>
           </timeline-badge>
-          <a class="timeline-link" href="javascript:void(0);" ng-click="apiEventsCtrl.selectEvent(eventTimeline)">
-            <timeline-panel ng-class="{active : apiEventsCtrl.isEventSelected(eventTimeline)}" class="{{eventTimeline.badgeClass}}">
+            <timeline-panel class="{{eventTimeline.badgeClass}}">
               <timeline-heading>
-                <h5>{{eventTimeline.title | uppercase}}
-                  <ng-md-icon class="api-events-icon" ng-show="apiEventsCtrl.eventToCompareRequired && !apiEventsCtrl.isEventSelected(eventTimeline)" icon="crop_din" style="fill: rgba(0,0,0,0.54);"></ng-md-icon>
-                  <ng-md-icon class="api-events-icon" ng-show="apiEventsCtrl.isEventSelectedForComparaison(eventTimeline)" icon="compare" style="fill: green;"></ng-md-icon>
-                </h5>
+                <h5>{{eventTimeline.title | uppercase}}</h5>
                 <p ng-if="eventTimeline.when">
                   <small class="text-muted ng-binding"><i class="glyphicon glyphicon-time"></i>{{eventTimeline.when | date : 'medium'}} by {{eventTimeline.username}}</small>
                 </p>
-                <p ng-if="eventTimeline.isCurrentAPI">
-                  <small class="text-muted ng-binding"><i class="glyphicon glyphicon-info-sign"></i>The current undeployed API</small>
-                </p>
               </timeline-heading>
             </timeline-panel>
-          </a>
         </timeline-event>
       </div>
     </timeline>
   </div>
-  
-  <div class="api-events-content" flex="80">
-    <div ng-show="apiEventsCtrl.eventsSelected.length == 1" class="api-events-content-descriptor">
-      <md-card style="box-shadow:none;">
-        <div ng-show="apiEventsCtrl.eventsTimeline.length > 1" style="border: 1px solid #e7eaec; border-bottom: 0;" layout-padding>
-          <md-button ng-show="apiEventsCtrl.eventsSelected[0].isCurrentAPI" class="md-raised" ng-click="apiEventsCtrl.$scope.$parent.apiCtrl.showDeployAPIConfirm($event, apiEventsCtrl.api)">Deploy</md-button>
-          <md-button ng-show="(!apiEventsCtrl.eventsSelected[0].isCurrentAPI && apiEventsCtrl.eventsSelected[0].event.id != apiEventsCtrl.events[0].id) || (!apiEventsCtrl.eventsSelected[0].isCurrentAPI && apiEventsCtrl.eventsTimeline[0].isCurrentAPI)" class="md-raised" ng-click="apiEventsCtrl.showRollbackAPIConfirm($event, apiEventsCtrl.eventSelectedPayload)" ng-click="apiEventsCtrl.showRollbackAPIConfirm($event, apiEventsCtrl.eventSelectedPayload)">Rollback</md-button>
-          <md-button ng-show="apiEventsCtrl.events.length > 0 && apiEventsCtrl.eventsSelected[0].event.id != apiEventsCtrl.events[0].id" class="md-raised" ng-click="apiEventsCtrl.diffWithMaster()">Diff with published API</md-button>
-          <md-button ng-show="apiEventsCtrl.eventsTimeline.length > 1" class="md-raised" ng-click="apiEventsCtrl.enableDiff()">Diff</md-button>
-        </div>
-        <md-card-content style="background-color: #fff; border: 1px solid #e7eaec; padding-top: 0px">
-          <pre ng-show="!apiEventsCtrl.diffMode" class="api-events-definition">{{apiEventsCtrl.eventSelectedPayloadDefinition | json }}</pre>
-          <pre ng-show="apiEventsCtrl.diffMode" class="api-events-definition" gravitee-diff old-value="apiEventsCtrl.left" new-value="apiEventsCtrl.right"></pre>
-        </md-card-content>
-      </md-card>    
-    </div>
-  </div>
-  
 </div>
