+++ src/app/api/admin/history/_apiHistory.scss
@@ -1,0 +1,68 @@
+.api-history-timeline {
+  overflow-y: auto;
+  height: calc(100vh - 200px);
+}
+
+.api-history-timeline .timeline li timeline-badge {
+  z-index: 0 !important;
+}
+
+.api-history-timeline .timeline li timeline-panel {
+  padding: 7px !important;
+}
+
+.api-history-timeline .timeline li timeline-panel:before {
+  top: 25px !important;
+}
+
+.api-history-timeline .timeline li timeline-panel:after {
+  top: 26px !important;
+}
+
+.api-history-timeline .timeline timeline-panel.active {
+  background-color: #ededed;
+}
+
+.api-history-timeline .timeline timeline-panel.active:after {
+  border-right: none;
+}
+
+.api-history-timeline .timeline .timeline-link {
+  color: #333;
+}
+
+.api-history-timeline .timeline:before {
+  position: absolute;
+  width: 1px;
+  left: 6%;
+}
+
+.api-history-timeline .timeline li timeline-badge {
+  box-shadow: none;
+  width: 20px;
+  height: 20px;
+  line-height: 20px;
+  font-size: 1em;
+  left: 13%;
+  top: 29px;
+}
+
+.api-history-timeline .timeline li.timeline-inverted timeline-panel {
+  float:left;
+  left: 18%;
+}
+
+.api-history-timeline ul.timeline li timeline-panel {
+  width: 80%;
+}
+
+pre.api-history-definition {
+  background-color: transparent !important;
+  border: none !important;
+}
+
+.api-history-icon {
+  position: absolute;
+  margin-top: 5px;
+  right: 9px;
+}
